backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio==5.11.4
starlette==0.38.2
pydantic==2.9.2
pydantic-settings==2.5.2
motor==3.6.0
redis==5.0.8
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
numpy==2.1.1
scikit-learn==1.5.2
xgboost==2.1.1
torch==2.3.1
transformers==4.44.2
